- Child class - added an add child from console method - in progress, not done here
- MyApp -
put 2 of the test children on to the waiting list for testing purposes

- Person -
added a simpler toString that didn't include Username, for printing phonelist

- UserInterface -
Started work on Employee menu, with submenus found below.

Added waiting list method, that prints a list with all the kids on the waiting list. Should work as intended

Added telephone list method, that prints either employees or parents numbers to screen. Should work as intended