- Child class - added default constructor for making an empty object added an add child from console method - in progress, not done here added back in the childID in toStringSimpleChild, to make deleting children easier
- Parent -
added toStringSimplePrintParent to print a simple version of parent info, with parent id and phone number

- MyApp -
put 2 of the test children on to the waiting list for testing purposes

- Person -
added a simpler toString that didn't include Username, for printing phonelist

- UserInterface -
Started work on Employee menu, with submenus found below.

Added waiting list method, that prints a list with all the kids on the waiting list. Should work as intended

Added adding and removing kids from the waiting list.

Added telephone list method, that prints either employees or parents numbers to screen. Should work as intended

added children's submenu that lets user add, remove or see full list of children.

Added addChildFromConsole method, that takes the user through the proccess of making a new kid (this also includes asking if the parent already exists or not, and in that case making one for the new child)

Fixed some formatting and visual stuff to make it look more tight.